Fix: Correct status update logic in UbkApplicationChangeStatusService
The `changeStatus` method in `UbkApplicationChangeStatusServiceImp` was
incorrectly setting the application's status to the old status ID instead
of the new one. This prevented the status from being updated correctly.

This commit corrects the logic to use the new status ID from the request
DTO. A new test suite, `UbkApplicationChangeStatusServiceImpTest`, has
been added to verify the correctness of the new implementation.

+++ ubk/src/main/java/kg/mlsp/ubk/service/UbkCalculateGMDService.java
@@ -1,7 +1,5 @@
 package kg.mlsp.ubk.service;
-import java.math.BigDecimal;
-
 public interface UbkCalculateGMDService {
-    BigDecimal calculateGMD(Integer applicationId);
+    Double calculateGMD(Integer applicationId);
 
 }
